One more fix for #7 #8

+++ psl-core/src/main/java/edu/umd/cs/psl/util/database/Queries.java
@@ -143,7 +143,7 @@ public class Queries {
 						break;
 					case Long:
 						if (rawArgs[i] instanceof Long)
-							args[i] = new LongAttribute((Integer) rawArgs[i]);
+							args[i] = new LongAttribute((Long) rawArgs[i]);
 						else
 							throw new IllegalArgumentException("Could not convert raw arg " + i + " to Long.");
 					case Date:
